feat: Added lesson viewer to vm instance page
The lesson viewer is a dedicated panel to provide a guided training that
is meant to be executed in the virtual machine.

fix: The virtual machine is now able to autoscale

The virtual machine VNC viewer is now able to scale to match the screen
viewport as well as addition of fullscreen and button to reboot machine
using the VNC connection.

feat: Added a spinner on non-final status

All statuses that are not stopped or running will now show a loader to
indicate that the platform is reconciling the issue to reach final status.